Restructures server details page
Improves the server details page by restructuring it with tabs for general information, certificates, and web console.

This change introduces a new `ServerDetails` component acting as a container with navigation tabs.
The previous `ServerDetails` is now `GeneralServerDetails` and becomes the default content for the main tab.

Also, this commit sets up nested routes using React Router to manage the different sections within the server details page.
