+++ package.json
@@ -31,13 +31,12 @@
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
   },
-  "scripts": {
-    "start": "cross-env REACT_APP_VERSION=$npm_package_version react-scripts start",
+"scripts": {
+  "start": "cross-env REACT_APP_VERSION=$npm_package_version REACT_APP_PROXY_TARGET=http://localhost:5581 react-scripts start",
     "build": "cross-env REACT_APP_VERSION=$npm_package_version react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
-  "proxy": "http://localhost:5581",
   "eslintConfig": {
     "extends": [
       "react-app",
